fixed some spec/code with it

+++ app/controllers/articles_controller.rb
@@ -35,10 +35,7 @@ class ArticlesController < ApplicationController
   end
 
   def update
-    unless @article.user == current_user
-      flash[:danger] = "You can only edit your own article."
-      redirect_to root_path
-    else
+    if @article.user == current_user
       if @article.update(article_params)
         flash[:success] = 'Article has been updated'
         redirect_to @article
@@ -46,13 +43,21 @@ class ArticlesController < ApplicationController
         flash.now[:danger] = 'Article has not been updated'
         render :edit
       end
+    else
+      flash[:danger] = "You can only edit your own article."
+      redirect_to root_path
     end
   end
 
   def destroy
-    if @article.destroy
-      flash[:success] = "Article has been deleted."
-      redirect_to articles_path
+    if @article.user == current_user
+      if @article.destroy
+        flash[:success] = "Article has been deleted."
+        redirect_to articles_path
+      end
+    else
+      flash.now[:alert] = 'You can only delete your own article.'
+      redirect_to root_path
     end
   end
